pyproject.toml
[project]
name = "simvue-cli"
version = "1.0.2"
description = "Command Line Interface for interaction with a Simvue v3 server"
authors = [
    {name = "Simvue Development Team", email = "info@simvue.io"}
]
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.10,<3.14"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Operating System :: Unix",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Monitoring",
    "Topic :: Utilities",
    "Typing :: Typed"
]
keywords = [
    "tracking",
    "monitoring",
    "metrics",
    "alerting",
    "metrics-gathering"
]
dependencies = [
    "click-option-group>=0.5.7",
    "click>=8.1.8",
    "simvue>=2.1.1",
    "click-log>=0.4.0",
    "click-params>=0.5.0",
    "toml>=0.10.2",
    "tabulate>=0.9.0",
    "regex>=2024.11.6",
    "requests>=2.32.3",
    "humanfriendly>=10.0",
    "ukaea-multiparser>=1.0.4",
    "pyyaml>=6.0.2",
]

[project.urls]
homepage = "https://simvue.io"
repository = "https://github.com/simvue-io/simvue-cli"
documentation = "https://docs.simvue.io"

[project.scripts]
simvue = "simvue_cli.cli:simvue"

[tool.ruff]
lint.extend-select = ["C901"]
lint.mccabe.max-complexity = 11
extend-exclude = ["tests"]

[tool.pytest.ini_options]
addopts = "--no-cov"
testpaths = [
    "tests"
]
markers = [
  "unix: UNIX only tests"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.5",
]
lint = [
    "ruff>=0.11.2",
]

[tool.mypy]
ignore_missing_imports = true
